// src/app/features/borrowed-stock-query/models/borrowed-reply-query-response.model.ts
export interface BorrowedReplyQueryResponse {
  no: string;
  brdate: string;
  stock: string;
  dseq: string;
  rpdate: string;
  rptype: string;
  rptypenm: string;
  Cqty: string;
  Dbfee: string;
  Brfee: string;
  Fee: string;
  Cdnamt: string;
  Dnint: string;
  Dninttax: string;
  netamt: string;
}